fix: resolve Unity C# compile errors in port stubs
- PigController.Yaw setter now public for BotPigAI

- DimensionFlipper.Is3D setter now public for GameManager

- AudioManager.dspTime typed as double

- LaserWeapon removes invalid RaycastHit.collider assignment

- PoopBomb uses local Vector3 copy to mutate Velocity

+++ unity-port/WhenPigsCanFlyUnity/Assets/Scripts/AudioManager.cs
@@ -50,7 +50,7 @@ namespace WhenPigsCanFly
             if (!Playing) return;
             float bpm = 90f + (Level - 1) * 8f;
             float interval = 60f / bpm;
-            float time = AudioSettings.dspTime;
+            double time = AudioSettings.dspTime;
 
             while (nextBeatTime < time + 0.12f)
             {
